requirements.txt
absl-py==0.13.0
aiofiles==0.5.0
aiohttp==3.7.4.post0
aniso8601==7.0.0
anykeystore==0.2
apex==0.9.10.dev0
appdirs==1.4.3
apturl==0.5.2
arabic-reshaper==2.1.3
asgiref==3.3.4
astunparse==1.6.3
async-exit-stack==1.0.1
async-generator==1.10
async-timeout==3.0.1
asyncio==3.4.3
attrs==19.3.0
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
cached-property==1.5.1
cachetools==4.2.2
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==2.0.3
chrome-gnome-shell==0.0.0
Click==7.0
colorama==0.4.3
command-not-found==0.3
coolname==1.1.0
cryptacular==1.5.5
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.2
dnspython==2.1.0
docker==4.1.0
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.12.0
email-validator==1.1.3
entrypoints==0.3
fastapi==0.66.0
fasteners==0.14.1
filelock==3.0.12
flatbuffers==1.12
future==0.18.2
gast==0.4.0
gensim==4.0.1
google-auth==1.33.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
googletrans==3.1.0a0
graphene==2.1.9
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet==1.1.0
grpcio==1.34.1
h11==0.9.0
h2==3.2.0
h5py==3.1.0
hpack==3.0.0
hstspreload==2021.7.5
httpcore==0.9.1
httplib2==0.14.0
httptools==0.1.2
httpx==0.13.3
hupper==1.10.3
hyperframe==5.2.0
idna==2.8
importlib-metadata==1.5.0
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
keyring==18.0.1
kiwisolver==1.3.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libretranslatepy==2.1.1
lockfile==0.12.2
louis==3.12.0
lxml==4.6.3
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.3.4
MarkupSafe==1.1.0
matplotlib==3.4.2
monotonic==1.5
more-itertools==4.2.0
multidict==5.1.0
netifaces==0.10.4
networkx==2.5.1
node2vec==0.4.3
numpy==1.19.5
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.3.0
orjson==3.6.0
pandas==1.3.0
paramiko==2.6.0
PasteDeploy==2.1.1
pbkdf2==1.3
pexpect==4.6.0
Pillow==7.0.0
plaster==1.0
plaster-pastedeploy==0.7
promise==2.3
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycups==1.9.73
pydantic==1.8.2
Pygments==2.3.1
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.7
PyQt5==5.14.1
pyramid==2.0
pyramid-mailer==0.15.1
pyRFC3339==1.1
pyrsistent==0.15.5
python-apt==2.0.0+ubuntu0.20.4.4
python-bidi==0.4.2
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-dotenv==0.18.0
python-Levenshtein==0.12.2
python-multipart==0.0.5
python3-openid==3.2.0
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
repoze.sendmail==4.4.1
requests==2.26.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rfc3986==1.5.0
rsa==4.7.2
runx==0.0.11
Rx==1.6.1
scikit-learn==0.24.2
scipy==1.7.0
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
simplejson==3.16.0
sip==4.19.21
six==1.15.0
sklearn==0.0
smart-open==5.1.0
sniffio==1.2.0
SQLAlchemy==1.4.21
sqlparse==0.4.1
starlette==0.14.2
systemd-python==234
tabulate==0.8.9
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.4
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
texttable==1.6.2
threadpoolctl==2.2.0
tqdm==4.61.2
transaction==3.0.1
translate==3.6.1
translationstring==1.4
typing-extensions==3.7.4.3
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
ujson==4.0.2
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
uvicorn==0.13.4
uvloop==0.15.3
velruse==1.1.1
venusian==3.0.0
virtualenv==20.0.17
wadllib==1.3.3
watchgod==0.7
WebOb==1.8.7
websocket-client==0.53.0
websockets==8.1
Werkzeug==2.0.1
wrapt==1.12.1
WTForms==2.3.3
wtforms-recaptcha==0.3.2
xkit==0.0.0
yarl==1.6.3
zipp==1.0.0
zope.deprecation==4.4.0
zope.interface==5.4.0
zope.sqlalchemy==1.5
